fix: use updated defaults for TCP check degraded/max response time (#306)
The maximum values were reduced at the backend level and the defaults were
changed as well. This change syncs both values to current values used by the
backend.

This change is technically backwards incompatible, but even the current
version is broken right now if the user relies on default values.