feat: rename heartbeat and TCP check resources to monitors
The old resource types are still available, and have not been deprecated.
If we deprecate them, users will start seeing warnings every time they
deploy their resources. In the Plugin SDK however there is no convenient
way to move a resource to a new type, so the steps to rename a resource
are a bit convoluted:
  1. Figure out the ID of the affected check
  2. Rename the resource (name, type) in the .tf file
  3. Use the import command or preferably the import block to import the
     existing resource into the new resource type
  4. Use terraform state rm to remove the old resource from TF state

With the framework, a single "moved" block would do the job (well, we'd
have to have extra code for it in the provider though).